Conflicts:
	externals/javelin
	src/__celerity_resource_map__.php
	src/__phutil_library_map__.php
	src/applications/directory/controller/PhabricatorDirectoryController.php
	src/applications/directory/controller/PhabricatorDirectoryMainController.php
	src/applications/maniphest/controller/ManiphestTaskDetailController.php
	src/applications/metamta/storage/PhabricatorMetaMTAReceivedMail.php
	src/applications/phid/PhabricatorObjectHandle.php
	src/infrastructure/markup/PhabricatorMarkupEngine.php
	src/infrastructure/util/PhabricatorSlug.php
	src/view/form/control/PhabricatorRemarkupControl.php
	src/view/layout/headsup/AphrontHeadsupView.php
	src/view/page/menu/PhabricatorMainMenuView.php
	webroot/rsrc/css/aphront/context-bar.css
	webroot/rsrc/css/aphront/form-view.css
	webroot/rsrc/css/aphront/headsup-action-list-view.css
	webroot/rsrc/css/aphront/headsup.css
	webroot/rsrc/css/aphront/panel-view.css
	webroot/rsrc/css/aphront/side-nav-view.css
	webroot/rsrc/css/aphront/table-view.css
	webroot/rsrc/css/aphront/tokenizer.css
	webroot/rsrc/css/aphront/transaction.css
	webroot/rsrc/css/application/base/main-menu-view.css
	webroot/rsrc/css/application/base/notification-menu.css
	webroot/rsrc/css/application/base/standard-page-view.css
	webroot/rsrc/css/application/directory/phabricator-app-buttons.css
	webroot/rsrc/css/application/directory/phabricator-jump-nav.css
	webroot/rsrc/css/application/feed/feed.css
	webroot/rsrc/css/application/maniphest/task-summary.css
	webroot/rsrc/css/application/maniphest/transaction-detail.css
	webroot/rsrc/css/autosprite.css
	webroot/rsrc/css/core/buttons.css
	webroot/rsrc/css/core/core.css

+++ src/applications/files/storage/PhabricatorFile.php
@@ -5,6 +5,9 @@ final class PhabricatorFile extends PhabricatorFileDAO
 
   const STORAGE_FORMAT_RAW  = 'raw';
 
+  const METADATA_IMAGE_WIDTH  = 'width';
+  const METADATA_IMAGE_HEIGHT = 'height';
+
   protected $phid;
   protected $name;
   protected $mimeType;
@@ -12,14 +15,20 @@ final class PhabricatorFile extends PhabricatorFileDAO
   protected $authorPHID;
   protected $secretKey;
   protected $contentHash;
+  protected $metadata = array();
 
   protected $storageEngine;
   protected $storageFormat;
   protected $storageHandle;
 
+  protected $ttl;
+
   public function getConfiguration() {
     return array(
       self::CONFIG_AUX_PHID => true,
+      self::CONFIG_SERIALIZATION => array(
+        'metadata' => self::SERIALIZATION_JSON,
+      ),
     ) + parent::getConfiguration();
   }
 
@@ -114,7 +123,7 @@ final class PhabricatorFile extends PhabricatorFileDAO
     $file = id(new PhabricatorFile())->loadOneWhere(
       'name = %s AND contentHash = %s LIMIT 1',
       self::normalizeFileName(idx($params, 'name')),
-      PhabricatorHash::digest($data));
+      self::hashFileContent($data));
 
     if (!$file) {
       $unguarded = AphrontWriteGuard::beginScopedUnguardedWrites();
@@ -125,11 +134,58 @@ final class PhabricatorFile extends PhabricatorFileDAO
     return $file;
   }
 
+  public static function newFileFromContentHash($hash, $params) {
 
-  public static function newFromFileData($data, array $params = array()) {
+    // Check to see if a file with same contentHash exist
+    $file = id(new PhabricatorFile())->loadOneWhere(
+      'contentHash = %s LIMIT 1', $hash);
+
+    if ($file) {
+      // copy storageEngine, storageHandle, storageFormat
+      $copy_of_storage_engine = $file->getStorageEngine();
+      $copy_of_storage_handle = $file->getStorageHandle();
+      $copy_of_storage_format = $file->getStorageFormat();
+      $copy_of_byteSize = $file->getByteSize();
+      $copy_of_mimeType = $file->getMimeType();
+
+      $file_name = idx($params, 'name');
+      $file_name = self::normalizeFileName($file_name);
+      $file_ttl = idx($params, 'ttl');
+      $authorPHID = idx($params, 'authorPHID');
+
+      $new_file = new  PhabricatorFile();
+
+      $new_file->setName($file_name);
+      $new_file->setByteSize($copy_of_byteSize);
+      $new_file->setAuthorPHID($authorPHID);
+      $new_file->setTtl($file_ttl);
+
+      $new_file->setContentHash($hash);
+      $new_file->setStorageEngine($copy_of_storage_engine);
+      $new_file->setStorageHandle($copy_of_storage_handle);
+      $new_file->setStorageFormat($copy_of_storage_format);
+      $new_file->setMimeType($copy_of_mimeType);
+
+      $new_file->save();
+
+      return $new_file;
+    }
+
+    return $file;
+  }
+
+  private static function buildFromFileData($data, array $params = array()) {
     $selector = PhabricatorEnv::newObjectFromConfig('storage.engine-selector');
 
-    $engines = $selector->selectStorageEngines($data, $params);
+    if (isset($params['storageEngines'])) {
+      $engines = $params['storageEngines'];
+    } else {
+      $selector = PhabricatorEnv::newObjectFromConfig(
+        'storage.engine-selector');
+      $engines = $selector->selectStorageEngines($data, $params);
+    }
+
+    assert_instances_of($engines, 'PhabricatorFileStorageEngine');
     if (!$engines) {
       throw new Exception("No valid storage engines are available!");
     }
@@ -171,6 +227,7 @@ final class PhabricatorFile extends PhabricatorFileDAO
 
     $file_name = idx($params, 'name');
     $file_name = self::normalizeFileName($file_name);
+    $file_ttl = idx($params, 'ttl');
 
     // If for whatever reason, authorPHID isn't passed as a param
     // (always the case with newFromFileDownload()), store a ''
@@ -179,7 +236,8 @@ final class PhabricatorFile extends PhabricatorFileDAO
     $file->setName($file_name);
     $file->setByteSize(strlen($data));
     $file->setAuthorPHID($authorPHID);
-    $file->setContentHash(PhabricatorHash::digest($data));
+    $file->setTtl($file_ttl);
+    $file->setContentHash(self::hashFileContent($data));
 
     $file->setStorageEngine($engine_identifier);
     $file->setStorageHandle($data_handle);
@@ -196,9 +254,26 @@ final class PhabricatorFile extends PhabricatorFileDAO
       $file->setMimeType(Filesystem::getMimeType($tmp));
     }
 
+    try {
+      $file->updateDimensions(false);
+    } catch (Exception $ex) {
+      // Do nothing
+    }
+
     $file->save();
 
     return $file;
+  }
+
+  public static function newFromFileData($data, array $params = array()) {
+    $hash = self::hashFileContent($data);
+    $file = self::newFileFromContentHash($hash, $params);
+
+    if ($file) {
+      return $file;
+    }
+
+    return self::buildFromFileData($data, $params);
   }
 
   public function migrateToEngine(PhabricatorFileStorageEngine $engine) {
@@ -258,7 +333,12 @@ final class PhabricatorFile extends PhabricatorFileDAO
   }
 
 
-  public static function newFromFileDownload($uri, $name) {
+  public static function newFromFileDownload($uri, array $params = array()) {
+    // Make sure we're allowed to make a request first
+    if (!PhabricatorEnv::getEnvConfig('security.allow-outbound-http')) {
+      throw new Exception("Outbound HTTP requests are disabled!");
+    }
+
     $uri = new PhutilURI($uri);
 
     $protocol = $uri->getProtocol();
@@ -273,12 +353,15 @@ final class PhabricatorFile extends PhabricatorFileDAO
 
     $timeout = 5;
 
-    $file_data = HTTPSFuture::loadContent($uri, $timeout);
-    if ($file_data === false) {
-      return null;
-    }
+    list($file_data) = id(new HTTPSFuture($uri))
+        ->setTimeout($timeout)
+        ->resolvex();
 
-    return self::newFromFileData($file_data, array('name' => $name));
+    $params = $params + array(
+      'name' => basename($uri),
+    );
+
+    return self::newFromFileData($file_data, $params);
   }
 
   public static function normalizeFileName($file_name) {
@@ -286,13 +369,34 @@ final class PhabricatorFile extends PhabricatorFileDAO
   }
 
   public function delete() {
-    $engine = $this->instantiateStorageEngine();
+    // delete all records of this file in transformedfile
+    $trans_files = id(new PhabricatorTransformedFile())->loadAllWhere(
+      'TransformedPHID = %s', $this->getPHID());
 
+    $this->openTransaction();
+    foreach ($trans_files as $trans_file) {
+      $trans_file->delete();
+    }
     $ret = parent::delete();
+    $this->saveTransaction();
 
-    $engine->deleteFile($this->getStorageHandle());
+    // Check to see if other files are using storage
+    $other_file = id(new PhabricatorFile())->loadAllWhere(
+      'storageEngine = %s AND storageHandle = %s AND
+      storageFormat = %s AND id != %d LIMIT 1', $this->getStorageEngine(),
+      $this->getStorageHandle(), $this->getStorageFormat(),
+      $this->getID());
 
+    // If this is the only file using the storage, delete storage
+    if (count($other_file) == 0) {
+      $engine = $this->instantiateStorageEngine();
+      $engine->deleteFile($this->getStorageHandle());
+    }
     return $ret;
+  }
+
+  public static function hashFileContent($data) {
+    return sha1($data);
   }
 
   public function loadFileData() {
@@ -353,6 +457,12 @@ final class PhabricatorFile extends PhabricatorFileDAO
     return PhabricatorEnv::getCDNURI($path);
   }
 
+  public function getPreview140URI() {
+    $path = '/file/xform/preview-140/'.$this->getPHID().'/'
+      .$this->getSecretKey().'/';
+    return PhabricatorEnv::getCDNURI($path);
+  }
+
   public function getPreview220URI() {
     $path = '/file/xform/preview-220/'.$this->getPHID().'/'
       .$this->getSecretKey().'/';
@@ -361,6 +471,12 @@ final class PhabricatorFile extends PhabricatorFileDAO
 
   public function getThumb220x165URI() {
     $path = '/file/xform/thumb-220x165/'.$this->getPHID().'/'
+      .$this->getSecretKey().'/';
+    return PhabricatorEnv::getCDNURI($path);
+  }
+
+  public function getThumb280x210URI() {
+    $path = '/file/xform/thumb-280x210/'.$this->getPHID().'/'
       .$this->getSecretKey().'/';
     return PhabricatorEnv::getCDNURI($path);
   }
@@ -470,6 +586,12 @@ final class PhabricatorFile extends PhabricatorFileDAO
     return idx($mime_map, $mime_type);
   }
 
+  public function getDisplayIconForMimeType() {
+    $mime_map = PhabricatorEnv::getEnvConfig('files.icon-mime-types');
+    $mime_type = $this->getMimeType();
+    return idx($mime_map, $mime_type, 'docs_file');
+  }
+
   public function validateSecretKey($key) {
     return ($key == $this->getSecretKey());
   }
@@ -483,6 +605,51 @@ final class PhabricatorFile extends PhabricatorFileDAO
 
   public function generateSecretKey() {
     return Filesystem::readRandomCharacters(20);
+  }
+
+  public function updateDimensions($save = true) {
+    if (!$this->isViewableImage()) {
+      throw new Exception(
+        "This file is not a viewable image.");
+    }
+
+    if (!function_exists("imagecreatefromstring")) {
+      throw new Exception(
+        "Cannot retrieve image information.");
+    }
+
+    $data = $this->loadFileData();
+
+    $img = imagecreatefromstring($data);
+    if ($img === false) {
+      throw new Exception(
+        "Error when decoding image.");
+    }
+
+    $this->metadata[self::METADATA_IMAGE_WIDTH] = imagesx($img);
+    $this->metadata[self::METADATA_IMAGE_HEIGHT] = imagesy($img);
+
+    if ($save) {
+      $this->save();
+    }
+
+    return $this;
+  }
+
+  public static function getMetadataName($metadata) {
+    switch ($metadata) {
+      case self::METADATA_IMAGE_WIDTH:
+        $name = pht('Width');
+        break;
+      case self::METADATA_IMAGE_HEIGHT:
+        $name = pht('Height');
+        break;
+      default:
+        $name = ucfirst($metadata);
+        break;
+    }
+
+    return $name;
   }
